+++ crates/orchestrator/src/jobs/state_update_job/mod.rs
@@ -109,11 +109,12 @@ impl Job for StateUpdateJob {
             block_numbers = block_numbers.into_iter().filter(|&block| block >= last_failed_block).collect::<Vec<u64>>();
         }
 
+        let mut nonce = config.settlement_client().get_nonce().await.map_err(|e| JobError::Other(OtherError(e)))?;
+
         let mut sent_tx_hashes: Vec<String> = Vec::with_capacity(block_numbers.len());
         for block_no in block_numbers.iter() {
             let snos = self.fetch_snos_for_block(*block_no).await;
-
-            let tx_hash = self.update_state_for_block(config, *block_no, snos).await.map_err(|e| {
+            let tx_hash = self.update_state_for_block(config, *block_no, snos, nonce).await.map_err(|e| {
                 job.metadata.insert(JOB_METADATA_STATE_UPDATE_LAST_FAILED_BLOCK_NO.into(), block_no.to_string());
 
                 self.insert_attempts_into_metadata(job, &attempt_no, &sent_tx_hashes);
@@ -123,6 +124,7 @@ impl Job for StateUpdateJob {
                 )))
             })?;
             sent_tx_hashes.push(tx_hash);
+            nonce += 1;
         }
 
         self.insert_attempts_into_metadata(job, &attempt_no, &sent_tx_hashes);
@@ -265,6 +267,7 @@ impl StateUpdateJob {
         config: &Config,
         block_no: u64,
         snos: StarknetOsOutput,
+        nonce: u64,
     ) -> Result<String, JobError> {
         let settlement_client = config.settlement_client();
         let last_tx_hash_executed = if snos.use_kzg_da == Felt252::ZERO {
@@ -272,9 +275,10 @@ impl StateUpdateJob {
         } else if snos.use_kzg_da == Felt252::ONE {
             let blob_data = fetch_blob_data_for_block(block_no).await.map_err(|e| JobError::Other(OtherError(e)))?;
 
+            // Fetching nonce before the transaction is run
             // Sending update_state transaction from the settlement client
             settlement_client
-                .update_state_with_blobs(vec![], blob_data)
+                .update_state_with_blobs(vec![], blob_data, nonce)
                 .await
                 .map_err(|e| JobError::Other(OtherError(e)))?
         } else {